+++ src/api-mocks/msw-server.js
@@ -1,0 +1,6 @@
+import {setupServer} from 'msw/node';
+
+// set up API mock server
+const mswServer = setupServer();
+
+export default mswServer;
