+++ src/components/LanguageSelection/LanguageSelection.stories.mdx
@@ -7,10 +7,10 @@ import {useState} from 'react';
 import {IntlProvider} from 'react-intl';
 
 import {ConfigContext} from 'Context';
+import {BASE_URL} from 'api-mocks';
 import ErrorBoundary from 'components/ErrorBoundary';
 import {I18NContext} from 'i18n';
 import {ConfigDecorator} from 'story-utils/decorators';
-import {BASE_URL} from 'story-utils/mocks';
 
 import {LanguageSelection, LanguageSelectionDisplay} from '.';
 import {
